+++ docs/CHANGELOG.md
@@ -11,11 +11,13 @@ All notable changes to this project will be documented in this file. The format 
 - Regular expressions for extracting filesystem property substrings
 - A plugin for phys2bids to convert physiological data
 - Option to install extra packages, such as phys2bids
+- Added a multi-echo deface function `medeface` that uses the same defacemask for all echo-images
 
 ### Changed
 - Plugins should now have a `is_sourcefile` and a `get_attribute` function and have a simpler API (-> DataSource class)
 - The intricate filtering of the `nrfiles` property by the other filesystem properties has been removed and is now a pure/unfiltered file-system property
 - The `<<SourceFilePath>>` keyword has been replaced by a more flexible filepath regular expression
+- The dcm2bidsmap and the dcm2niix2bids plugins have been merged
 
 ## [3.6.3] - 2021-06-14